refactor: splitted e2e ot individual dart files
